The migration to the DI container is complete.
- Migration to a DI container completed
- Optimization
- Improved data display in the console
- Improved message sending

+++ Bot/Models/Users/UserData.cs
@@ -5,9 +5,9 @@
         public required string Id { get; set; }
         public required string Language { get; set; }
         public required string Name { get; set; }
-        public int? Balance { get; set; }
-        public int? BalanceFloat { get; set; }
-        public int? TotalMessages { get; set; }
+        public long? Balance { get; set; }
+        public long? BalanceFloat { get; set; }
+        public long? TotalMessages { get; set; }
         public bool? IsBanned { get; set; }
         public bool? Ignored { get; set; }
         public bool? IsModerator { get; set; }
